Split scratch.jsx into clearer parts
Refactor to split scratch.jsx now that it's responsibilities have grown.

I've kept scratch.jsx as responsible for initializing the page and setting up authentication. ScratchEditor is responsible for rendering the Scratch editor component and setting up the handlers it needs.

I've also extracted WrappedScratchGui as it makes the tests a lot easier to set up when you don't need to mock the process of setting up the higher order component.
